更新数据库结构

+++ dbscript/update_credit.sql
@@ -19,15 +19,15 @@ VALUES
 
 DROP TABLE IF EXISTS `sa_goods_category`;
 CREATE TABLE `sa_goods_category` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `pid` int(11) DEFAULT NULL COMMENT '父分类ID',
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `pid` BIGINT DEFAULT NULL COMMENT '父分类ID',
   `title` varchar(100) DEFAULT NULL COMMENT '分类名称',
   `short` varchar(20) DEFAULT NULL COMMENT '分类简称',
   `name` varchar(50) DEFAULT NULL COMMENT '分类别名',
   `icon` varchar(100) DEFAULT NULL COMMENT '图标',
   `image` varchar(100) DEFAULT NULL COMMENT '大图',
   `props` TEXT COMMENT '绑定属性',
-  `sort` int(11) DEFAULT NULL COMMENT '排序',
+  `sort` int DEFAULT NULL COMMENT '排序',
   `keywords` varchar(255) DEFAULT NULL COMMENT '分类关键词',
   `description` varchar(255) DEFAULT NULL COMMENT '分类描述',
   PRIMARY KEY (`id`)
@@ -35,8 +35,8 @@ CREATE TABLE `sa_goods_category` (
 
 DROP TABLE IF EXISTS `sa_goods`;
 CREATE TABLE `sa_goods` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `cate_id` int(11) DEFAULT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `cate_id` BIGINT DEFAULT NULL,
   `title` varchar(150) DEFAULT NULL,
   `vice_title` varchar(200) DEFAULT NULL,
   `unit` varchar(10) DEFAULT NULL,
@@ -46,50 +46,50 @@ CREATE TABLE `sa_goods` (
   `price` DECIMAL(10,2) DEFAULT 0 COMMENT '兑换价格',
   `market_price` DECIMAL(10,2) DEFAULT 0 COMMENT '市场价格',
   `content` text,
-  `create_time` int(11) DEFAULT '0',
-  `update_time` int(11) DEFAULT '0',
+  `create_time` BIGINT DEFAULT '0',
+  `update_time` BIGINT DEFAULT '0',
   `levels` varchar(100) DEFAULT '',
-  `storage` int(11) DEFAULT '0',
-  `limit` int(11) DEFAULT '0' COMMENT '兑换次数',
-  `sort` int(11) DEFAULT '0',
-  `sale` int(11) DEFAULT '0' COMMENT '总销量',
-  `type` tinyint(4) DEFAULT '1',
-  `status` tinyint(4) NOT NULL DEFAULT '1',
+  `storage` int DEFAULT '0',
+  `limit` int DEFAULT '0' COMMENT '兑换次数',
+  `sort` int DEFAULT '0',
+  `sale` int DEFAULT '0' COMMENT '总销量',
+  `type` tinyint DEFAULT '1',
+  `status` tinyint NOT NULL DEFAULT '1',
   PRIMARY KEY (`id`),
   KEY `cate_id` (`cate_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 DROP TABLE IF EXISTS `sa_goods_images`;
 CREATE TABLE `sa_goods_images` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `title` varchar(100) DEFAULT NULL,
   `description` varchar(250) DEFAULT NULL,
   `image` varchar(150) DEFAULT NULL,
-  `goods_id` int(11) DEFAULT NULL,
-  `sort` INT(11) NOT NULL DEFAULT '0',
+  `goods_id` BIGINT DEFAULT NULL,
+  `sort` INT NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `goods_id` (`goods_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 DROP TABLE IF EXISTS `sa_credit_order`;
 CREATE TABLE `sa_credit_order` (
-  `order_id` INT NOT NULL AUTO_INCREMENT,
+  `order_id` BIGINT NOT NULL AUTO_INCREMENT,
   `order_no` VARCHAR(30) NULL,
-  `member_id` INT NULL,
+  `member_id` BIGINT NULL,
   `paycredit` DECIMAL(10,2) NULL DEFAULT 0,
   `payamount` DECIMAL(10,2) NULL DEFAULT 0,
-  `create_time` INT NULL DEFAULT 0,
+  `create_time` BIGINT NULL DEFAULT 0,
   `pay_type` VARCHAR(20) NULL COMMENT '付款方式',
-  `pay_time` INT NULL DEFAULT 0,
-  `deliver_time` INT NULL DEFAULT 0,
-  `confirm_time` INT NULL DEFAULT 0,
-  `comment_time` INT NULL DEFAULT 0,
-  `refund_time` INT NULL DEFAULT 0,
-  `cancel_time` INT NULL DEFAULT 0,
+  `pay_time` BIGINT NULL DEFAULT 0,
+  `deliver_time` BIGINT NULL DEFAULT 0,
+  `confirm_time` BIGINT NULL DEFAULT 0,
+  `comment_time` BIGINT NULL DEFAULT 0,
+  `refund_time` BIGINT NULL DEFAULT 0,
+  `cancel_time` BIGINT NULL DEFAULT 0,
   `status` TINYINT NULL DEFAULT 0 COMMENT '订单状态',
-  `delete_time` INT NULL DEFAULT 0 COMMENT '删除状态',
+  `delete_time` BIGINT NULL DEFAULT 0 COMMENT '删除状态',
   `remark` VARCHAR(250) NULL,
-  `address_id` INT NULL,
+  `address_id` BIGINT NULL,
   `receive_name` VARCHAR(45) NULL,
   `mobile` VARCHAR(45) NULL,
   `province` VARCHAR(50) NULL,
@@ -100,21 +100,21 @@ CREATE TABLE `sa_credit_order` (
   `express_no` VARCHAR(100) NULL,
   `express_code` VARCHAR(20) NULL,
   `express_data` TEXT NULL,
-  `express_time` INT NULL DEFAULT 0,
+  `express_time` BIGINT NULL DEFAULT 0,
   PRIMARY KEY (`order_id`)
 )ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 DROP TABLE IF EXISTS `sa_credit_order_goods`;
 CREATE TABLE `sa_credit_order_goods` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `order_id` INT(11) DEFAULT '0',
-  `goods_id` INT(11) DEFAULT '0',
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `order_id` BIGINT DEFAULT '0',
+  `goods_id` BIGINT DEFAULT '0',
   `goods_title` varchar(100) DEFAULT NULL,
   `goods_image` varchar(150) DEFAULT NULL,
   `goods_orig_price` DECIMAL(10,2) DEFAULT NULL,
   `goods_price` DECIMAL(10,2) DEFAULT NULL,
-  `count` int(11) DEFAULT NULL,
-  `sort` INT(11) NOT NULL DEFAULT '0',
+  `count` int DEFAULT NULL,
+  `sort` INT NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `goods_id` (`goods_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
